Bug fix: NullPointerException during registration of new user.
The problem was that the method validate in RegisterServlet tested the values of user.date.format and user.time.format for all user, including the new ones. Since new users are not able to enter these values, they are always null. A test for user ID = -1 was added to correct this behaviour.